Some refactoring to better comply with SoC principle
