+++ src/hermes/commands/__init__.py
@@ -1,3 +1,13 @@
 # SPDX-FileCopyrightText: 2022 German Aerospace Center (DLR)
 #
 # SPDX-License-Identifier: Apache-2.0
+
+# SPDX-FileContributor: Michael Meinel
+
+from hermes.commands.base import HermesHelpCommand
+from hermes.commands.clean.base import HermesCleanCommand
+from hermes.commands.curate.base import HermesCurateCommand
+from hermes.commands.harvest.base import HermesHarvestCommand
+from hermes.commands.process.base import HermesProcessCommand
+from hermes.commands.deposit.base import HermesDepositCommand
+from hermes.commands.postprocess.base import HermesPostprocessCommand
