+++ src/hermes/commands/curate/base.py
@@ -1,0 +1,7 @@
+from hermes.commands.base import HermesCommand
+
+
+class HermesCurateCommand(HermesCommand):
+    """ Curate the unified metadata before deposition. """
+
+    command_name = "curate"
