+++ src/hermes/commands/harvest/base.py
@@ -1,0 +1,15 @@
+# SPDX-FileCopyrightText: 2022 German Aerospace Center (DLR)
+#
+# SPDX-License-Identifier: Apache-2.0
+
+# SPDX-FileContributor: Michael Meinel
+
+from hermes.commands.base import HermesCommand
+from hermes.settings import HarvestSettings
+
+
+class HermesHarvestCommand(HermesCommand):
+    """ Harvest metadata from configured sources. """
+
+    command_name = "harvest"
+    settings_class = HarvestSettings
