+++ src/hermes/commands/process/base.py
@@ -1,0 +1,13 @@
+# SPDX-FileCopyrightText: 2022 German Aerospace Center (DLR)
+#
+# SPDX-License-Identifier: Apache-2.0
+
+# SPDX-FileContributor: Michael Meinel
+
+from hermes.commands.base import HermesCommand
+
+
+class HermesProcessCommand(HermesCommand):
+    """ Process the collected metadata into a common dataset. """
+
+    command_name = "process"
